% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node-centrality.R
\name{cent_degree}
\alias{cent_degree}
\alias{cent_degree.igraph}
\alias{cent_degree.network}
\title{Degree Centrality}
\usage{
cent_degree(net, direction = c("in", "out", "all"),
  loops = net_has_loops(net), use_edge_weights = FALSE,
  node_index = node_seq(net), allow_bipartite = FALSE, ...)

\method{cent_degree}{igraph}(net, direction = c("in", "out", "all"),
  loops = net_has_loops(net), use_edge_weights = FALSE,
  node_index = node_seq(net), allow_bipartite = FALSE, ...)

\method{cent_degree}{network}(net, direction = c("in", "out", "all"),
  loops = net_has_loops(net), use_edge_weights = FALSE,
  node_index = node_seq(net), allow_bipartite = FALSE, ...)
}
\arguments{
\item{net}{net}

\item{direction}{direction of ties}

\item{loops}{whether to include loops}

\item{node_index}{index of target nodes}

\item{allow_bipartite}{allow bipartite}

\item{...}{named arguments passed on to lower-level functions}
}
\value{
\code{numeric} \code{vector}
}
\description{
Degree Centrality
}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
