% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edge_attr_exists}
\alias{edge_attr_exists}
\title{Does an edge attribute exist?}
\usage{
edge_attr_exists(.net, .edge_attr)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}

\item{.edge_attr}{\code{character} scalar corresponding to the name of a edge attribute in \code{.net}.}
}
\value{
\code{logical} scalar
}
\description{
Does an edge attribute exist?
}
\examples{
edge_attr_exists(.net = florence_business_ig, .edge_attr = "edge_type")

florence_business_nw \%>\% 
  edge_attr_exists(.edge_attr = "fake attr")

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
