% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edges.R
\name{edge_count}
\alias{edge_count}
\alias{edge_count.igraph}
\alias{edge_count.network}
\title{Count the number of edges in a network.}
\usage{
edge_count(.net)

\method{edge_count}{igraph}(.net)

\method{edge_count}{network}(.net)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}
}
\value{
\code{numeric} scalar
}
\description{
Count the number of edges in a network.
}
\examples{
edge_count(.net = florence_business_ig)

florence_business_nw \%>\% 
  edge_count()
  
florence_combo_ig \%>\% 
  edge_count() 

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
