% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-jemaah_islamiyah.R
\docType{data}
\name{jemaah_islamiyah_ig}
\alias{jemaah_islamiyah_ig}
\alias{jemaah_islamiyah_nw}
\title{Koschade's Jemaah Islamiyah (2002 Bali Bombings)}
\format{\itemize{
\item jemaah_islamiyah_ig: \code{igraph}
\item jemaah_islamiyah_nw: \code{network}
}}
\source{
\url{https://sites.google.com/site/ucinetsoftware/datasets/covert-networks/jemaahislamiyahkoschade}
}
\usage{
jemaah_islamiyah_ig

jemaah_islamiyah_nw
}
\description{
\code{jemmah_islamiyah} is an adaptation of the \href{https://sites.google.com/site/ucinetsoftware/}{UCINET}
dataset "Jemaah Islamiyah Koschade". \cr
}
\details{
From the \href{https://sites.google.com/site/ucinetsoftware/datasets/covert-networks/jemaahislamiyahkoschade}{UCINET documentation}: \cr

Jemaah Islamiyah cell that was responsible for the Bali bombings
in 2002 – i.e. should tally with the other Bali bombing dataset from JJATT. The
recording of the interaction of the cell began following the  meeting in the Hotel
Harem in Denpasar on October 6, when the group was considered to go ‘operationally
covert’, and concluded when the majority of the group had left Bali before the
implementation of the operation on October 11, 2002.
}
\examples{
jemaah_islamiyah_ig

jemaah_islamiyah_nw

}
\references{
Koschade, Stuart (2006) A Social Network Analysis of Jemaah Islamiyah: The
Applications to Counter-Terrorism and Intelligence. Studies in Conflict and Terrorism
Vol. 29(6):pp. 559-575

International Crisis Group. 2003 ‘Jemaah Islamiyah in South East Asia:
Damaged but Still Dangerous’
}
\keyword{datasets}
