% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net_has_isolates}
\alias{net_has_isolates}
\alias{net_has_isolates.igraph}
\alias{net_has_isolates.network}
\title{Does a network have any isolate nodes?}
\usage{
net_has_isolates(.net)

\method{net_has_isolates}{igraph}(.net)

\method{net_has_isolates}{network}(.net)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}
}
\value{
\code{logical} scalar
}
\description{
Does a network have any isolate nodes?
}
\examples{
net_has_isolates(.net = florence_business_ig)

florence_business_nw \%>\% 
  net_has_isolates()

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
