% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net_has_loops}
\alias{net_has_loops}
\alias{net_has_loops.igraph}
\alias{net_has_loops.network}
\title{Does a network have any loop edges?}
\usage{
net_has_loops(.net)

\method{net_has_loops}{igraph}(.net)

\method{net_has_loops}{network}(.net)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}
}
\value{
\code{logical} scalar
}
\description{
Does a network have any loop edges?
}
\examples{
net_has_loops(.net = florence_business_ig)

florence_business_nw \%>\% 
  net_has_loops()

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
