% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net_is_bipartite}
\alias{net_is_bipartite}
\alias{net_is_bipartite.igraph}
\alias{net_is_bipartite.network}
\title{Is a network bipartite?}
\usage{
net_is_bipartite(.net)

\method{net_is_bipartite}{igraph}(.net)

\method{net_is_bipartite}{network}(.net)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}
}
\value{
\code{logical} scalar
}
\description{
Is a network bipartite?
}
\examples{
net_is_bipartite(.net = florence_business_ig)

florence_business_nw \%>\% 
  net_is_bipartite()

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
