% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net_is_multiplex}
\alias{net_is_multiplex}
\alias{net_is_multiplex.igraph}
\alias{net_is_multiplex.network}
\title{Is a network multiplex?}
\usage{
net_is_multiplex(.net)

\method{net_is_multiplex}{igraph}(.net)

\method{net_is_multiplex}{network}(.net)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}
}
\value{
\code{logical} scalar
}
\description{
Is a network multiplex?
}
\examples{
net_is_multiplex(.net = florence_business_ig)

florence_business_nw \%>\% 
  net_is_multiplex()

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
