% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.R
\name{net_is_weighted}
\alias{net_is_weighted}
\title{Are a network's edges weighted?}
\usage{
net_is_weighted(.net)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}
}
\value{
\code{logical} scalar
}
\description{
Are a network's edges weighted?
}
\examples{
net_is_weighted(.net = florence_business_ig)

florence_business_nw \%>\% 
  net_is_weighted()

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
