% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{node_attr_exists}
\alias{node_attr_exists}
\title{Does a node attribute exist?}
\usage{
node_attr_exists(.net, .node_attr)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}

\item{.node_attr}{\code{character} scalar corresponding to the name of a node attribute in \code{.net}.}
}
\value{
\code{logical} scalar
}
\description{
Does a node attribute exist?
}
\examples{
node_attr_exists(.net = florence_business_ig, .node_attr = "wealth")

florence_business_nw \%>\% 
  node_attr_exists(.node_attr = "priors")
  
florence_business_ig \%>\% 
  node_attr_exists(.node_attr = "fake attr")

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
