% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\name{node_get_names}
\alias{node_get_names}
\alias{node_name_attr}
\title{Extract node names.}
\usage{
node_get_names(.net)

node_name_attr(.net)
}
\arguments{
\item{.net}{network of class \code{igraph} or \code{network}.}
}
\value{
\itemize{
\item \code{node_get_names()}
\itemize{
\item \code{character} \code{vector}
}
}

\itemize{
\item \code{node_name_attr()}:
\itemize{
\item \code{character} scalar
}
}
}
\description{
Extract node names.
}
\details{
\itemize{
\item \code{node_get_names()}
\itemize{
\item Uses \code{node_name_attr()} to determine node names.
}
}

\itemize{
\item \code{node_name_attr()}
\itemize{
\item If \code{.net} is an \code{igraph} object, \code{node_get_names()} returns the node (vertex)
attribute \code{"name"}.
\itemize{
\item If \code{"name"} is not a node attribute, \code{node_get_names()} returns \code{node_seq(.net)}.
}
\item If \code{.net} is a \code{network} object, \code{node_get_names()} returns the node (vertex)
attribute \code{"vertex.names"}.
}
}
}
\examples{
node_get_names(.net = florence_business_ig)

florence_business_nw \%>\% 
  node_get_names()

node_name_attr(.net = florence_business_ig)

florence_business_nw \%>\% 
  node_name_attr()

}
\author{
Brendan Knapp \email{brendan.knapp@nps.edu}
}
